Conflicts:
	htroot/ConfigLanguage_p.html
	htroot/ConfigLanguage_p.java
	source/net/yacy/data/Translator.java
	source/net/yacy/utils/translation/CreateTranslationMasters.java

+++ source/net/yacy/htroot/ConfigLanguage_p.java
@@ -1,10 +1,41 @@
+//ConfigLanguage_p.java
+//-----------------------
+//part of YACY
+//(C) by Michael Peter Christen; mc@yacy.net
+//first published on http://www.anomic.de
+//Frankfurt, Germany, 2004
+
+//This File is contributed by Alexander Schier
+//
+//$LastChangedDate$
+//$LastChangedRevision$
+//$LastChangedBy$
+//
+//This program is free software; you can redistribute it and/or modify
+//it under the terms of the GNU General Public License as published by
+//the Free Software Foundation; either version 2 of the License, or
+//(at your option) any later version.
+
+//This program is distributed in the hope that it will be useful,
+//but WITHOUT ANY WARRANTY; without even the implied warranty of
+//MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//GNU General Public License for more details.
+
+//You should have received a copy of the GNU General Public License
+//along with this program; if not, write to the Free Software
+//Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+
+//You must compile this file with
+//javac -classpath .:../Classes Blacklist_p.java
+//if the shell's current path is HTROOT
+
 package net.yacy.htroot;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
-import java.net.URISyntaxException;
+import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
@@ -12,6 +43,8 @@ import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+
+import com.google.common.io.Files;
 
 import net.yacy.cora.document.id.DigestURL;
 import net.yacy.cora.protocol.ClientIdentification;
@@ -23,6 +56,7 @@ import net.yacy.kelondro.util.ResourceUtils;
 import net.yacy.search.Switchboard;
 import net.yacy.server.serverObjects;
 import net.yacy.server.serverSwitch;
+import net.yacy.utils.translation.TranslatorXliff;
 
 
 public class ConfigLanguage_p {
@@ -31,49 +65,60 @@ public class ConfigLanguage_p {
 
         final serverObjects prop = new serverObjects();
         Switchboard sb = (Switchboard) env;
-        final URL langPathURL = sb.getAppFileOrDefaultResource("locale.source", "/locales/");
+        URL customLangPathURL = null;
+        File customLangDir = sb.getDataPath("locale.source", "DATA/LOCALE").getAbsoluteFile();
+		try {
+			customLangPathURL = customLangDir.toURI().toURL();
+		} catch (MalformedURLException ignored) {
+		}
 
         //Fallback
         //prop.put("currentlang", ""); //is done by Translationtemplate
         prop.put("status", "0");//nothing
 
-        List<URL> langFiles = Translator.langFiles(langPathURL);
-        if (langFiles == null) {
+        if(customLangPathURL == null) {
+        	return prop;
+        }
+        
+        List<URL> langURLs = Translator.langFiles(customLangPathURL);
+        if (langURLs == null) {
             return prop;
         }
 
         if (post != null) {
-            final String selectedLanguage = post.get("language");
+            final String selectedLangKey = post.get("language");
+            URL selectedLangFile = null;
+            if(selectedLangKey != null && !selectedLangKey.isEmpty()) {
+            	try {
+            		selectedLangFile = new URL(customLangPathURL, selectedLangKey  + ".lng");
+            	} catch (MalformedURLException ignored) {
+            	}
+            }
 
             // store this call as api call
-            ((Switchboard) env).tables.recordAPICall(post, "ConfigLanguage_p.html", WorkTables.TABLE_API_TYPE_CONFIGURATION, "language settings: " + selectedLanguage);
+            ((Switchboard) env).tables.recordAPICall(post, "ConfigLanguage_p.html", WorkTables.TABLE_API_TYPE_CONFIGURATION, "language settings: " + selectedLangKey);
 
             //change language
-            if (post.containsKey("use_button") && selectedLanguage != null){
+            if (post.containsKey("use_button") && selectedLangFile != null){
                 /* Only change language if filename is contained in list of filesnames
                  * read from the language directory. This is very important to prevent
                  * directory traversal attacks!
                  */
-                if (langFiles.contains(selectedLanguage) || selectedLanguage.startsWith("default")) {
-                    Translator.changeLang(env, langPathURL, selectedLanguage);
+                if (langURLs.contains(selectedLangFile) || "default".equals(selectedLangKey)) {
+                    Translator.changeLang(env, selectedLangFile, selectedLangKey);
                 }
 
                 //delete language file
-            } else if (post.containsKey("delete")) {
-
-                /* Only delete file if filename is contained in list of filesnames
-                 * read from the language directory. This is very important to prevent
-                 * directory traversal attacks!
-                 */
-            	File langDir = null;
-            	try {
-					langDir = new File(langPathURL.toURI());
-				} catch (URISyntaxException ignored) {
+            } else if (post.containsKey("delete") && !"default".equals(selectedLangKey)) {
+				/*
+				 * Only delete file if filename is contained in list of
+				 * filesnames read from the language directory. This is very
+				 * important to prevent directory traversal attacks!
+				 */
+				if (langURLs.contains(selectedLangFile)) {
+					File langfile = new File(customLangDir, selectedLangKey + ".lng");
+					FileUtils.deletedelete(langfile);
 				}
-                if (langFiles.contains(selectedLanguage) && langDir != null) {
-                    final File langfile= new File(langDir, selectedLanguage);
-                    FileUtils.deletedelete(langfile);
-                }
 
                 //load language file from URL
             } else if (post.containsKey("url")){
@@ -83,23 +128,33 @@ public class ConfigLanguage_p {
                     final DigestURL u = new DigestURL(url);
                     it = FileUtils.strings(u.get(ClientIdentification.yacyInternetCrawlerAgent, null, null));
                 	try {
-                    	final File langFile = new File(new File(langPathURL.toURI()), url.substring(url.lastIndexOf('/'), url.length()));
+                		customLangDir.mkdirs();
+                    	File langFile = new File(customLangDir, url.substring(url.lastIndexOf('/'), url.length()));
                     	final OutputStreamWriter bw = new OutputStreamWriter(new FileOutputStream(langFile), StandardCharsets.UTF_8.name());
 
                     	while (it.hasNext()) {
                         	bw.write(it.next() + "\n");
                     	}
                     	bw.close();
+                    	
+                        // convert downloaded xliff to internal lng file
+                        final String ext = Files.getFileExtension(langFile.getName());
+						if (ext.equalsIgnoreCase("xlf") || ext.equalsIgnoreCase("xliff")) {
+							TranslatorXliff tx = new TranslatorXliff();
+							Map<String, Map<String, String>> lng = TranslatorXliff
+									.loadTranslationsListsFromXliff(langFile);
+							langFile = new File(customLangDir,
+									Files.getNameWithoutExtension(langFile.getName()) + ".lng");
+							tx.saveAsLngFile(null, langFile, lng);
+						}
+                    	
                 		if (post.containsKey("use_lang") && "on".equals(post.get("use_lang"))) {
-                    		Translator.changeLang(env, langPathURL, url.substring(url.lastIndexOf('/'), url.length()));
+                    		Translator.changeLang(env, selectedLangFile, Files.getNameWithoutExtension(langFile.getName()));
                 		}
                 	} catch(final IOException e) {
                     	prop.put("status", "2");//error saving the language file
                     	return prop;
-                	} catch (URISyntaxException e) {
-                    	prop.put("status", "2");//error saving the language file
-                    	return prop;
-					}
+                	}
                 } catch(final IOException e) {
                     prop.put("status", "1");//unable to get url
                     prop.put("status_url", url);
@@ -108,8 +163,8 @@ public class ConfigLanguage_p {
         }
 
         //re-read language files
-        langFiles = Translator.langFiles(langPathURL);
-        Collections.sort(langFiles, new Comparator<URL>() {
+        langURLs = Translator.langFiles(customLangPathURL);
+        Collections.sort(langURLs, new Comparator<URL>() {
 
 			@Override
 			public int compare(URL o1, URL o2) {
@@ -117,23 +172,23 @@ public class ConfigLanguage_p {
 				if(o1 != null && o2 != null) {
 					res = o1.toExternalForm().compareTo(o2.toExternalForm());
 				}
-				return 0;
+				return res;
 			}
 		});
         final Map<String, String> langNames = Translator.langMap(env);
 
         //virtual entry
-        prop.put("langlist_0_file", "default");
+        prop.put("langlist_0_key", "default");
         prop.put("langlist_0_name", ((langNames.get("default") == null) ? "default" : langNames.get("default")));
         prop.put("langlist_0_selected", "selected=\"selected\"");
 
         int count = 0;
-        for (final URL langFileURL : langFiles) {
+        for (final URL langFileURL : langURLs) {
             //+1 because of the virtual entry "default" at top
         	String langFile = ResourceUtils.getFileName(langFileURL);
             final String langKey = langFile.substring(0, langFile.length() -4);
             final String langName = langNames.get(langKey);
-            prop.put("langlist_" + (count + 1) + "_file", langFile);
+            prop.put("langlist_" + (count + 1) + "_key", langKey);
             prop.put("langlist_" + (count + 1) + "_name", ((langName == null) ? langKey : langName));
 
             if(env.getConfig("locale.language", "default").equals(langKey)) {
